entity: Add an input-based jumping entity
Add InputJumper and DetectCollider to implement an entity that jumps
whenever it receives an specific input. Currently, this input is set
to Unity's 'Jump' axis, but it should eventually be changed.

DetectCollider should be used to detect when the object is touching
the floor. Then, the object with the InputJumper component may jump
if it detects the expected input.