level: Add Goal component
Add a Goal component that should be added to a scene as the stage's
objective. The Goal component signal the event to the root object, which
must have been previously configured.

Component AdvanceLevel implement a simple method of advancing to the next
stage. It shouldn't be used in a real scenario, but it works for
development.