ProjectSettings: Add custom physics layers
Add the custom physics layers:

- LevelController
- LevelComponent
- NoCollision
- PlayerController
- PlayerComponent
- Player

LevelController and LevelComponent only collide with each other, and so
does PlayerController and PlayerComponent. The *Component is intended for
tagging the entity receiving events (say, a moving platform) and the
*Controller is intended for the entity that send events (say, a force
field).

Player has the same collision as the Default layer, except it can't collide
with itself. NoCollision, as the name implies, can't collide with anything.