entity/MovingPlatform: Add public ForceCorrection factor
As far as I can tell, Unity's physics does not apply the force being
applied to an object to a colliding object. To work around that, the
MovingPlatform component manually send an event to touching entities
to apply its movement force onto these entities.

However, for some reason (mass? drag? something else?) if it were to
apply only its movement force, the touching entity would move slower
than the MovingPlayer and would get left behind. This was originally
fixed by multiplying the force by a constant factor.

Although this initially worked, this factor must change based on the
force being applied to the MovingPlatform. So, replace this constant
factor by a public property, which may be set per instance.

This works in most cases, specially for slow movements. For downward
movement, however, there are a few tricks and unsolvable issues.

When an entity is resting upon another, its vertical speed is set to
zero. So, if an entity is on top of a MovingPlatform going downward,
it will fall for a bit, but it will stop falling when it touches the
Platform. Since the MovingPlatform moves constantly, the entity will
end up hanging on the air before starting to fall again. This has an
additional side effect: the entity stops colliding with the Platform
and thus stop being applied the Platform's moving force.

I couldn't find any definitive solution for this issue. Limiting the
vertical speed of the Platform helps to keep the entity in place. In
addition to that, making the Platform's hitbox slightly taller makes
the Platform affect nearby entities, causing it to apply its falling
force to an entity that was previously on top of it. Even though the
entity briefly hangs in the air, it should come back in contact with
the MovingPlatform.