prefabs: Add the Root object
Create a prefab for the Root object, since it should be the same in most,
if not all, stages.

In Mystery Tower, my previous game, there's a main scene for controlling
the game and adding a background and the stage itself is loaded as a
sub-scene. In this scenario, it may be useful to have this sub-scene not
have any cameras.

To account for that, modify the Root's CameraHelper so it may instantiate
a camera, in case it doesn't detect any camera on the scene.