action: Refactor SetForce to use direction + force
When modifying multiple objects at once, it's easier to have the direction
and the force as independent properties. This way, even if various objects
were select and they had different directions, it would be possible to set
their force at the same time.

So, refactor SetForce's Force vector into a direction vector and a force.