prefabs: Slow down platform on MovingPlatformController contact
In its current implementation, the platforms in stage 05-across-the-gap are
quite likely to throw the player away when they come to a halt. One way to
solve that would be to have the BlockForTime and SetForce on different
objects and displaced a bit. This way, the Platform would first hit the
SetForce component, decreasing its speed, and then hit the BlockForTime
component.

Add the MovingPlatformController "meta component" for controlling both the
BlockForTime and SetForce components of a platform controller's at once.
This "meta component" also allows to separate both sub-components, to
automatically implement the aforementioned solution to throwing the player
away on halt.