prefabs: Add the Spring prefab
Add a physics-based Spring prefab. This object is 100% physics based, and
therefore can be a bit janky. The spring platform is moved by a Push event
and it's stopped by simply colliding against solid collider. This required
a new layer able to collide against the LevelController (for the invisible
colliders that make the platform stop) and against everything (so it may
physically interact with the player).

The spring could be made extra janky and fun by allowing the rigidbody to
rotate. However, most of than not this causes the platform to eventually
get thrown in a random direction.