level: Disable SetRoot if a Loader exists
When a stage is loaded from the loading scene, the loading scene will have
its own root object. Since each stage must have its own root object to work
inside the editor, the SetRoot component must still be used.

Disable the SetRoot component if running in release mode and if a Loader
exists when running in the Editor.