scene: Fix RecalibrateGamepad
Unity is quite finicky with inputs. When a scene is loaded, every input is
zeroed-out. Therefore, trying to train an input in this state will cause
lots of issues.

To work around this issue, wait until an input was pressed and request that
the user moves every axis around before training the gamepad. Also, modify
Input.TrainAxisStable() so it may train the default gamepad as well.