Add Flowing Fog particles
New Particles prefab to be used with Push Fields.

I don't know if this fog is consistent with the visual identity
of this game, we can make further changes if needed.

Compared to the current Push Field Fog, this one is less dense
and is worse for carrying colors (it seems to stay way "whiter"
and less saturated).