Add a "Simplified" version of the PushFieldAutomation Script
Oh boy, you'll kill me for this.

I said "Simplified", but this one is a completely new beast.

First of all, I didn't understand the necessity of the script knowing
the "direction" (it **could** be used to change the particle colors,
but wasn't). Just rotate the prefab...

Second, removed a lot of runtime object search in favor
of direct references, with this, the "programmer" knows
exactly which object is being used, also, no runtime search.

The new script also scales the particle **emission rate**
and the particle **speed**. :confia:

This is currently unused.