entity: Make PushField fade in/out on timed events
If a SwappingEntity component were to deactivate a PushField, the particles
would simply vanish where they were. Add an OnSwap event to stop the
particle emitters programmatically.

This has the catch that particles must then be emitted all over the area
and must live shortly. Otherwise, although no new particle would be emitted
after receiving an 'OnSwap(_, false)', the currently active particles would
continue simulating until they died... which is exactly the opposite of the
desired effect.

Modify PushFieldAutomationSimplified to make the emitter shape slightly
smaller than the push collider, based on the particle speed and on its
lifetime, allowing particles to spawn all over the desired area even
though short-lived. The default emission rate was increased in the
FlowingFog prefab to make this effect look better.

Lastly, modify the opacity of the fog texture by applying a color directly
on the ColorOverLifetime component of each PushField variant. The opacity
itself was also tweak (in general and directly on some variants) to make
the color more visible.