level: Add a background based on the level directory
Change the default skybox into a custom scene. This scene is loaded
automatically by the loader and must be named after the directory where the
level is located at. Therefore, for the level
"scenes/00-basic/00-first-steps.unity", the Loader tries to load the scene
"00-basic.unity" as the background.

The camera in this background scene should have a BGFollowCamera component,
which modifies the background camera based on the position of the main
camera in world space.