level: Fix rotating the camera back to neutral
Depending on the source and destination angles for resetting the camera
behind the player (e.g., if source were 270 and destination were 0), were
original calculation to check the orientation of the rotation would fail,
causing the camera to take the longer rotation.

Fix detecting whether this rotation should be clockwise or
counter-clockwise.