entity/InputJumper: Decrease fallback delay
Because of Unity's weird physics quirks, there's a brief delay after
issuing a jump and checking whether the player has actually left the floor.
To avoid issues if the player left and got back to the floor during this
delay, the game assumes that after a while the player must have left the
floor and waits until they get back to it to allow jumping again.

Decrease this delay and make it configurable (just in case).