level/Loader: Retry loading on error
Mystery Tower, which has a similar level/scene structure to this game, has
a weird bug (that I couldn't ever reproduce) in which the game skipped a
level and froze while loading the following level.

Bug demonstration: https://youtu.be/mnMlodhhpas?t=9604

Although the loader has been somewhat improved from Mystery Tower's, it
wold still stop working if any of the scenes failed to be loaded.

Capture any errors while loading a level and retry loading it, falling
back to the main menu if that doesn't work after a few attempts.